Final posted
Removed player vs player since it was causing to many issues trying to change anything i had to do everthing twice as much, plus it was hard to figure out how im suppoused to hide the ships and not and how the shot and posting of the shots or how many ships left. It was just to much so took that out, made it only going against a bot and simplified it.